#42 [Feat] CertificateConverter 수정
