+++ src/main/java/umc/forgrad/controller/CertificateController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import umc.forgrad.apipayload.ApiResponse;
+import umc.forgrad.converter.CertificateConverter;
 import umc.forgrad.domain.Certificate;
 import umc.forgrad.dto.CertificateRequestDto;
 import umc.forgrad.dto.CertificateResponseDto;
@@ -19,7 +20,7 @@ public class CertificateController {
     @PostMapping(value = "/plans/certifications/stuId/{stuId}")
     public ApiResponse<CertificateResponseDto> addCertificate(@RequestBody @Valid CertificateRequestDto certificateRequestDto) {
         Certificate certificate = certificateService.addCertificate(certificateRequestDto);
-        return ApiResponse.onSuccess(CertificateResponseDto.toAddResultDto(certificate));
+        return ApiResponse.onSuccess(CertificateConverter.toAddResultDto(certificate));
     }
