+++ src/main/java/umc/forgrad/service/CertificateService.java
@@ -3,6 +3,7 @@ package umc.forgrad.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import umc.forgrad.converter.CertificateConverter;
 import umc.forgrad.domain.Certificate;
 import umc.forgrad.dto.CertificateRequestDto;
 import umc.forgrad.repository.CertificateRepository;
@@ -14,7 +15,7 @@ public class CertificateService {
 
     @Transactional
     public Certificate addCertificate(CertificateRequestDto certificateRequestDto) {
-        return certificateRepository.save(certificateRequestDto.toEntity());
+        return certificateRepository.save(CertificateConverter.toCertificate(certificateRequestDto));
     }
 
     /*public Certificate viewCertificate(Long stuId) {
